+++ src/httphandler/constant.go
@@ -1,6 +1,7 @@
 package httphandler
 
 const (
+	SEND_PLAYER_ADMIN      = "A"
 	SEND_PLAYER_ID         = "0"
 	SEND_PLAYER_INIT       = "1"
 	SEND_PLAYER_UPDATE     = "2"
@@ -12,6 +13,8 @@ const (
 	SEND_SKY_COLOR          = "S"
 	SEND_SKY_COLOR_STANDARD = "S"
 	SEND_SKY_COLOR_FLASH    = "F"
+
+	SEND_LIGHT_STATE = "L"
 
 	RECEIVE_PLAYER_INIT   = "0"
 	RECEIVE_PLAYER_UPDATE = "1"