+++ src/player/create.go
@@ -1,5 +1,7 @@
 package player
 
 func Create() *Player {
-	return &Player{}
+	return &Player{
+		Size: 1,
+	}
 }